requirements.txt
# Digital Sarathi v2.0 Requirements

# Core Backend
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
pydantic[email]>=2.6.0
python-multipart>=0.0.9
python-dotenv>=1.0.1

# AI & LLM
langchain>=0.1.0
langchain-groq>=0.1.0
langgraph>=0.0.10
requests>=2.31.0
groq>=0.15.0

# Database
pymysql>=1.1.0
cryptography>=42.0.0

# Embeddings & Vector Store (if using RAG)
sentence-transformers>=2.5.0
faiss-cpu>=1.8.0
